chore(quill-charts): extract click routing into utils/resolve-clicked-bar-series
Move the pure `resolveClickedBarSeries` click-routing function out of
`BarChart.tsx` into a co-located `utils/resolve-clicked-bar-series.ts`. It was
already written as a pure, exported helper at the bottom of the component but
had no external importers; the move keeps the component focused on orchestration.

Pure refactor — no behavior or public API change; the existing DOM-level click
tests in `BarChart.test.tsx` cover the routing. Third of a few small PRs breaking
up the oversized `BarChart` component. Stacks on
`posthog-code/quill-barchart-2-config`.

Generated-By: PostHog Code
Task-Id: 8e22d6ce-b20c-4948-8774-4b148bce553b
